refactor(insights): move 5 dep-free leaves to @openloomi/insights (phase 6)
Extends the existing `@openloomi/insights` package with 5 more dep-free
leaves from `apps/web/lib/insights/`. These were already on the runtime
side of the boundary but hadn't been moved yet.

Moved:
- `constants`                — `DEBUG`, `EMAIL_TASK_LABEL`, `MAX_EMAIL_INSIGHTS`,
                                `CALENDAR_*`, `DEFAULT_CATEGORIES`, tag unions
- `grouping`                 — `normalizeMessagesInput`, `groupMessagesByChannel`,
                                `filterInsightsByGroup`, `mergeIMessageMessagesBySender`,
                                `estimateTokensForMessages`
- `platform-filter-config`   — `PLATFORM_FILTER_SUPPORT` + 4 helper functions
- `refresh-telemetry`        — `formatTimingError`, `shouldLogTimingEvent`
- `resolve-language`         — `resolveAgentLanguage`

Structural typing: `grouping.filterInsightsByGroup` is now generic
over `InsightWithGroups` instead of taking apps/web's `InsightData` via
`import("@/lib/ai/subagents/insights")`, and `platform-filter-config`
keys by a structural `PlatformId = string` instead of the
`@openloomi/contracts/integration-id` union. Both keep full type
compatibility for callers that pass the full `InsightData` or
`IntegrationId` shape.

Glue:
- `apps/web/lib/insights/{constants,grouping,platform-filter-config,
  refresh-telemetry,resolve-language}.ts` become one-line re-export shims.
- `apps/web/vitest.config.ts` gets explicit subpath aliases for the
  five new entry points.
- `packages/insights/src/index.ts` and `tsup.config.ts` updated to
  re-export the new modules.

The remaining 25+ files in `apps/web/lib/insights/` (processor.ts,
refresh.ts, embedding.ts, hebbian.ts, etc.) all pull in the Drizzle
DB, the AI subagent runtime, or the integrations glue — those are
out of scope for the dep-free leaf extraction and will be revisited
when their respective runtime services (`@openloomi/insights-service`,
HTTP daemon in a later phase) are scaffolded.

Verification: `tsc --noEmit` clean, `pnpm test` 2253 passed / 17
skipped / 0 failed, `pnpm --filter @openloomi/insights build` success.
