+++ apps/web/lib/insights/resolve-language.ts
@@ -1,21 +1,6 @@
-type LanguageSettings = {
-  language?: string | null;
-  languageAuto?: string | null;
-};
+// Phase 6 — re-export shim over `@openloomi/insights/resolve-language`.
+// The leaf module owns `resolveAgentLanguage` — picks the effective
+// agent-prompt language from explicit `language` / auto-learned
+// `languageAuto` settings. Pure utility, no external imports.
 
-/**
- * Resolve the effective language for agent prompts.
- *
- * Priority: explicit user setting (`language`) > auto-learned (`languageAuto`)
- * > none. The two write paths never overlap, so user changes and learning
- * cannot overwrite each other.
- */
-export function resolveAgentLanguage(
-  s: LanguageSettings | null | undefined,
-): string | null {
-  if (!s) return null;
-  const explicit = (s.language ?? "").trim();
-  if (explicit) return explicit;
-  const auto = (s.languageAuto ?? "").trim();
-  return auto || null;
-}
+export { resolveAgentLanguage } from "@openloomi/insights/resolve-language";
